+++ spec/checkers/column_presence_checker_spec.rb
@@ -90,5 +90,73 @@ RSpec.describe DatabaseConsistency::Checkers::ColumnPresenceChecker do
         end
       end
     end
+
+    if ActiveRecord::VERSION::MAJOR >= 5
+      required = { optional: false }
+      optional = { optional: true }
+    else
+      required = { required: true }
+      optional = { required: false }
+    end
+
+    context 'when null constraint is missing for the association key' do
+      before do
+        define_database_with_entity { |table| table.string :user_id }
+      end
+
+      let(:attribute) { :user }
+
+      context 'when `belongs_to` is required' do
+        let(:klass) { define_class { |klass| klass.belongs_to :user, **required } }
+
+        specify do
+          expect(checker.report).to have_attributes(
+            checker_name: 'ColumnPresenceChecker',
+            table_or_model_name: klass.name,
+            column_or_attribute_name: 'user',
+            status: :fail,
+            message: 'association foreign key column should be required in the database'
+          )
+        end
+      end
+
+      context 'when `belongs_to` is optional' do
+        let(:klass) { define_class { |klass| klass.belongs_to :user, **optional } }
+
+        specify do
+          expect(checker.report).to be_nil
+        end
+      end
+    end
+
+    context 'when null constraint is provided for the association key' do
+      let(:attribute) { :user }
+
+      before do
+        define_database_with_entity { |table| table.string :user_id, null: false }
+      end
+
+      context 'when `belongs_to` is required' do
+        let(:klass) { define_class { |klass| klass.belongs_to :user, **required } }
+
+        specify do
+          expect(checker.report).to have_attributes(
+            checker_name: 'ColumnPresenceChecker',
+            table_or_model_name: klass.name,
+            column_or_attribute_name: 'user',
+            status: :ok,
+            message: nil
+          )
+        end
+      end
+
+      context 'when `belongs_to` is optional' do
+        let(:klass) { define_class { |klass| klass.belongs_to :user, **optional } }
+
+        specify do
+          expect(checker.report).to be_nil
+        end
+      end
+    end
   end
 end